Avoid censoring curse words when they make up parts of different words.
For example, in this sentence:
Fish it out of the ocean.

The word 'shit' will match the last two letters of 'fish' the entire word 'it' and result in a false positive.

So when a match contains a space, we now check if it contains a space. If it contains a space, we will make sure it is also a word boundary (comes after a space or is the beginning of the line, and comes before a space or is the end of the line)